Move code for snapshot tags around a bit
Move code for generating snapshot tags and the fields used for said tags
to the snapshot_manager file. Also update the tag values:
* append "tag:" to all tag keys to comply with kopia CLI
* use a placeholder value as kopia tag comparisons don't work for empty
  values right now
